requirements.txt
torch
torchvision
transformers
evaluate
pillow
tqdm
